Ishaan and I created the method for finding a course section by CRN. I fixed some bugs that prevented the application from being tested and went back and changed some Ids that were longs to int (to match the pdf table). CourseSectionRepository has a method that was an issue which has been commented out because idk how to fix it but it has been marked to make sure others know not to uncomment it without fixing it.

+++ services/course-information/src/main/java/edu/uga/devdogs/course_information/Building/Building.java
@@ -17,7 +17,7 @@ public class Building implements Serializable {
 
     // serves as the id.
     @Id
-    private long buildingNumber;
+    private int buildingNumber;
 
     private String name;
 
@@ -54,7 +54,7 @@ public class Building implements Serializable {
      * Getters and Setters
      */
 
-    public long getBuildingNumber() {
+    public int getBuildingNumber() {
         return buildingNumber;
     }
